requirements.txt
customtkinter
pypdf
pdf2image
pytesseract
Pillow
img2pdf
